Clean up duplicate fixture code
